♻️ Refactor: 전체 값 가져오게 수정 전처리 및 폴더 구조 변경
