--- src/main/java/com/likelion/backendplus4/yakplus/scraper/drug/img/ApiDataDrugImgRepo.java
@@ -1,8 +1,0 @@
-package com.likelion.backendplus4.yakplus.scraper.drug.img;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ApiDataDrugImgRepo extends JpaRepository<ApiDataDrugImg, Long> {
-}
